docs: guide example readers and separate performance tools
