+++ benchmarks/workloads/compare.ts
@@ -1,0 +1,3 @@
+import { runPerformanceComparison } from "./workload";
+
+await runPerformanceComparison("short");
