+++ benchmarks/workloads/progress-long.ts
@@ -1,5 +1,5 @@
 import { Effect } from "effect";
-import { createPerformanceWorkload } from "./helpers/performance-workload";
+import { createPerformanceWorkload } from "./workload";
 
 const { progressRun } = createPerformanceWorkload("long");
 await Effect.runPromise(progressRun);
