+++ benchmarks/workloads/progress.ts
@@ -1,5 +1,5 @@
 import { Effect } from "effect";
-import { createPerformanceWorkload } from "./helpers/performance-workload";
+import { createPerformanceWorkload } from "./workload";
 
 const { progressRun } = createPerformanceWorkload("short");
 await Effect.runPromise(progressRun);
