+++ examples/edge-cases/failure.ts
@@ -1,5 +1,5 @@
 import { Effect } from "effect";
-import * as Progress from "../src";
+import * as Progress from "../../src";
 
 const program = Progress.all(
   Array.from({ length: 5 }).map((_, i) =>
