+++ examples/edge-cases/unknown-total.ts
@@ -1,5 +1,5 @@
 import { Effect } from "effect";
-import * as Progress from "../src";
+import * as Progress from "../../src";
 
 const wait = (millis: number) => Effect.sleep(`${millis} millis`);
