+++ examples/mixed-outcomes.ts
