--- examples/performanceComparisonLong.ts
@@ -1,3 +1,0 @@
-import { runPerformanceComparison } from "./helpers/performance-workload";
-
-await runPerformanceComparison("long");
