+++ package.json
@@ -40,9 +40,9 @@
     "format": "oxfmt --write .",
     "format:check": "oxfmt --check .",
     "check": "bun run typecheck && bun run lint && bun run format:check && knip",
-    "bench": "bun examples/benchmarks.ts",
-    "bench:render": "bun examples/renderBenchmarks.ts",
-    "perf": "bun --cpu-prof --cpu-prof-md examples/performance.ts"
+    "bench": "bun benchmarks/store-and-columns.ts",
+    "bench:render": "bun benchmarks/rendering.ts",
+    "perf": "bun --cpu-prof --cpu-prof-md benchmarks/workloads/progress.ts"
   },
   "dependencies": {
     "cli-spinners": "^3.4.0",
