+++ tools/ink/box-metrics.tsx
